RD-1591 Move some widget-related components to Stage.Shared.Widgets
