+++ app/components/PageContent.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import _ from 'lodash';
 import i18n from 'i18next';
-import WidgetsList from './WidgetsList';
+import { WidgetsList } from './shared/widgets';
 import { Confirm } from './basic';
 import AddWidget from '../containers/AddWidget';
 import './PageContent.css';
