+++ app/components/shared/widgets/WidgetsGrid.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Grid from './layout/Grid';
-import GridItem from './layout/GridItem';
+import Grid from '../../layout/Grid';
+import GridItem from '../../layout/GridItem';
 import Widget, { WidgetOwnProps } from './Widget';
 
 // NOTE: this cannot be a component since `Grid` expects only `GridItem`s as direct components
