+++ app/components/shared/widgets/index.ts
@@ -1,0 +1,5 @@
+import Widget from './Widget';
+import WidgetsGrid from './WidgetsGrid';
+import WidgetsList from './WidgetsList';
+
+export { Widget, WidgetsGrid, WidgetsList };
